requirements.txt
python-binance==1.0.16
pandas==2.0.3
